backend/requirements.txt
Django==5.2.0
djangorestframework==3.16.0
django-cors-headers==4.7.0
psycopg2-binary==2.9.10
Pillow==11.2.1
python-dotenv==1.1.0
django-environ==0.12.0